Model - update model slightly
By making detailPartsFitted a strict subset of detailParts, we can simplify all the crazy set code. First we just treat detailParts the same as the others to generate the tables, and then we do a pass over it to adjust the isFitted values based on whether or not the title is in the other field.

At the same time, match the model rename of lights/speakerFittings