Replace recordName field with recordID
There's no reason to ever store the bare recordName, it's just an
implementation detail of CloudKit. We can lookup the CKRecord.ID directly,
and that future-proofs us against additional fields being added.

We still keep a separate copy of the CKRecordZone.ID though because we're
going to want to handle "zone deleted" later on.