Consolidate object shenanigans out of the provider
This ends up being a lot of apparent changes, but the result is pretty
simple:

CloudStorable defines any type with a recordID that can create that ID
and create a CKRecord from its property values. Literally a type that is
storable in the cloud.

StorableManagedObject is any managed object with fields for recordID,
zoneID, and systemFields; as well as the functions to write from and to
a record.

We keep the two separate because then DetailPart just ends up being
a magic kind of CloudStorable that's not a StorableManagedObject.

NSManagedObject & StorableManagedObject get the CloudStorable impl for
free, along with an objectForRcordID() method to fetch or create.

NSManagedObject itself has a syncObjectFromRecord that uses a weird
protocol called StorableObjectTranslation to map the class type.