Model+Image - support deletion in batch requests too
Deleting the image accompanying the model can be tricky, it's fine when we have an NSManagedObject around because prepareDeletion() will get called - cleanup that case a little to explicitly delete rather than going through a property modifier, and ignore the "no such file" error entirely because such objects can exist in multiple contexts.

But it's harder when we issue a batch delete, we need to be able to run the same query but just check for image filenames, iterate over them, and delete them out - if possible avoiding actually doing a full fetch. So to accomplish this we add a class function to NSManagedObject and override that in Model; now we can do NSMO.willDeleteObjects(request, in: context) and know that for a model, that will result in the images being deleted.