CloudProvider - ensure we don't overlap fetch changes
We issue a fetch changes on app launch, and we also get remote notifications, and all of those can end up overlapping. While it's probably not the end of the world, and we'll probably end up with a consistent view, we'll end up repeating much of the same work over and over. So rather than do that, use a dispatch queue to line them up one after each other.