Model - replace ordered sets with an index property
The use of ordered relationships just doesn't end up playing well with NSFetchRequest and especially not NSFetchedResultsController, as well as being awkward to translate to the inverse relationships of CloudKit.

Switching to an index property is cleaner across all of them, at the slight extra cost of having to modify it.