IndexSortable - an abandoned approach to dealing with ordered sets
This (roughly) replicates the way that Z_FOK_table works in Core Data, using spaced out indexes so that most operations only need affect the one record that is being changed, and not needing adjustment of all.

In practice though, this doesn't work because for CloudKit we actually need a conflict in the case of placing two objects at the same index, and this wouldn't happen with this approach.