Split the Observer and Provider code apart
There's almost no code overlap between syncing from and syncing to the server, so split the code to make it simpler to read.

At the same time, rework the Observer code to

a) make it easier to reason about my making it a little more verbose in operation and

b) perform a two-stage process, using the local store as state rather than defaults.

This second part fixes the issue with handling zone changes and failed operation, since once a zone is marked dirty, it stays dirty until we finish fetches on it.