Date is the wrong type for storing Dates, use DateComponent
`Date` is an exact point in time, not a date on a calendar. This has meant that we've tried to use "midnight in the UTC timezone" as a method of storing a given day, but this results in all sorts of "5pm the day before" issues when rendered incorrectly.

A better approach is to store DateComponents, that way we can store {calendar:.gregorian, year:2018, month:12, day:25} and know that what we really intend is preserved, and it can be converted to a `Date` when necessary which will always be midnight in the local timezone.