Use a different strategy for ignoring sync changes
Right now the only reason we ignore changes from our own sync, and avoid loops, is that we only watch the viewContext for changes. This prevents us using any kind of background contexts for anything else.

A better strategy is to tag, using the provided userInfo dictionary, the specific contexts we don't want to loop on; which are the ones from the Observer applying cloud changes, and the Provider applying the updated system records.