Switch a read-only viewContext model
Use persistentContainer.viewContext as the managed object context for queries, and for the UI, but perform all editing and changes on background contexts that get merged back.

This gives us a very clean flow of data that we can reason about, and gives an edit flow where we can create a background context, update objects, and either save or discard as a single transaction.

Another benefit is that UI updates now only need to check for "Refreshed Objects".

There's a bunch of changes here:

We now pass down the NSPersistentContainer rather than NSManagedObjectContext, this is so we can create new background contexts when we need to.

 For TrainEditViewController:
 - we no longer set the train property, but add an editTrain method that creates a new background context and a copy of the object for editing
 - watch the background context, so make sure we reload the table on the main queue
 - since we're performing the save in the background, also make sure we dismiss the view, or show an alert, on the main queue

For TrainsViewController
 - watch the view context for changes, but refreshes only; make sure these are on the main queue (assertion to avoid C&P errors later)
 - save the result of a move on a background context, mostly just wrapping the code with an extra copy
 - store the changeIsUserDriven still, but we can't reset immediately after the save since a merge has to take place now, so restore after we see the change
 - since changeIsUserDriven is reset early, don't repeatedly check, instead rely on changeBlocks? being nil for other methods
 - call through to edit view with the method rather than property set