Model - rename `dateComponents` back to `date`
Even though the type is DateComponents, I'm using it to store a date (because `Date` is really a date-time). And unlike the `RawValue` case, I won't be providing a setter for using `Date` because that will be doing the wrong thing.

Rename it back, and rename `date` to be `dateAsDate` which better defines its role.