Remove NSBatchDeleteRequest
Since we weren't storing zones in the side-tables, we were relying on batch deletes cascading properly, as well as the RAII for the image files. This is all very shonky and an optimization for deletes which should be super-uncommon. Make everything much easier to reason about by switching to just deleting the object in the zone properly.