Price data entry fields
As we did with date, create Purchase.priceAsString and Purchase.valuationAsString, and replace the render in Purchase{Price,Valuation}TableViewCell with a call to these.

But then, unlike date right now, expose the formatter we use as Purchase.currencyFormatter with a few tweaks to correctly populate the currency symbols and codes by using Locale, and making it generate NSDecimalNumber.

Use this in Purchase{Price,Valuation}EditTableViewCell as a validator and converter.