Ignore calendar in DateComponents for now
Calendar includes a TimeZone, which means we'd end up putting the date in the entry timezone which is exactly what we were trying not to do!

So for now, don't put a calendar in the components, and always convert using our current Calendar (and TimeZone).

This works great for Gregorian in all timezones, but won't work for Hebrew date entry