Rework the added/edit/save/delete flow
Make it the responsibility of the presenting view controller to dismiss the modal, not the modal to dismiss itself.

Pass the completion handler for both add and edit methods, and have it return with an enum saying whether the result is that an object was saved (with the object), deleted, or if the changes were canceled.

For the saved and deleted cases, run the completion handler in the persistent container view context, but asynchronously. This gives the view context a chance to get the merge, and send out its own notifications, and thus update any views underneath the presented modal.

For the presenting list view, since the added case completion handler now runs after the table is updated, we can grab the indexPath out, select it, and perform the usual segue rather than needing a special "added one".

For the presenting edit view, avoid making changes to the view controller stack as a result of the notification, since we'd lose track of the modal - which is nice anyway as it means we isolate view stack and local view changes anyway, as we should.

Then for the presenting edit view we can dismiss the modal normally when the result was a cancel or save, but for a delete, we can be special - we can keep track of the real presenting view controller of the modal, pop our own view off the stack underneath it, and then perform an animated modal dismissal back to the parent on the stack.

In addition, handle the issue of popping the detail view controller off the navigation stack - by popping its own navigation controller off when we need to.