Use a read/write main queue context rather than a background context for Edit views
Keeping the viewConext read-only is a good pattern, and a separate context works well because we can live-edit the model and discard changes by simply not saving it. But a "background context" is the wrong choice for Edit views, we're supposed to always use the context's queue for any possible fetch operation, including property access, and can only update views from the main queue. That means a lot of bg.performAndWait{} which defeats the point of it being on the background at all.

A better pattern is to use a new main queue context, that writes changes back to the store