Redo train members handling
Simplify it to a single updateIndexes methods that will always fix gaps or duplicates, but also minimize the actual changes it makes when not necessary.

Change the move to be based on indexes rather than objects since we're always doing it in response to something that has indexes like a collection view or table view, also make it behave consistently with those where the given index is the same one that's returned rather than adjust for before/after changes.